requirements.txt
scikit-learn==1.1.3
scipy==1.10.1
seqeval==1.2.2
transformers==4.27.4
pytorch-crf==0.7.2
opencc==1.1.1
pandas==2.2.3
pytorch==2.5.1